+++ client/v2/autocli/common.go
@@ -1,10 +1,12 @@
 package autocli
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 	"google.golang.org/protobuf/reflect/protoreflect"
 	"sigs.k8s.io/yaml"
 
@@ -56,14 +58,21 @@ func (b *Builder) buildMethodCommandCommon(descriptor protoreflect.MethodDescrip
 		Version:      options.Version,
 	}
 
-	binder, err := b.AddMessageFlags(cmd.Context(), cmd.Flags(), inputType, options)
+	// the binder is only used for getting the number of args required
+	// it doesn't have access to the command context, so it isn't useful.
+	binderArgs, err := b.AddMessageFlags(context.Background(), &pflag.FlagSet{}, inputType, options)
 	if err != nil {
 		return nil, err
 	}
 
-	cmd.Args = binder.CobraArgs
+	cmd.Args = binderArgs.CobraArgs
 
 	cmd.RunE = func(cmd *cobra.Command, args []string) error {
+		binder, err := b.AddMessageFlags(cmd.Context(), cmd.Flags(), inputType, options)
+		if err != nil {
+			return err
+		}
+
 		input, err := binder.BuildMessage(args)
 		if err != nil {
 			return err